pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ev-bertopic"
version = "0.1.0"
description = "BERTopic extraction, rule-based refinement, and LLM refinement for EV Reddit data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "pandas",
  "scikit-learn",
  "matplotlib",
  "tqdm",
  "sentence-transformers",
  "bertopic",
  "umap-learn",
  "hdbscan",
  "google-genai",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
